requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
python-dotenv
psycopg2
catboost
xgboost
Flask
dill
mlflow==2.2.2

# -e .